Добавил админ команды по сбору участников, билетов, удаления юзера из лотереи, начала и остановки начисление билетов

+++ main.py
@@ -7,6 +7,7 @@ import models
 
 TOKEN = '5973304457:AAHGtaBx2VladoA7yt1S5H3IV7KDJoVD7z4'
 bot = telebot.TeleBot(TOKEN)
+stop = True
 
 @bot.message_handler(commands=['ref'])
 def send_referral_link(message):
@@ -41,6 +42,97 @@ def start(message):
                      reply_markup=check
                      )
 
+@bot.message_handler(commands=['reset'])
+def drop(message):
+    user_id = message.from_user.id
+    user_username = message.from_user.username
+    verification = config.is_admin(user_username, user_id)
+    if verification[0]:
+        models.clear()
+        bot.send_message(message.chat.id, "Все пользователи были удалены")
+
+
+@bot.message_handler(commands=['reset_tickets'])
+def drop(message):
+    user_id = message.from_user.id
+    user_username = message.from_user.username
+    verification = config.is_admin(user_username, user_id)
+    if verification[0]:
+        models.db.connect()
+        for user in models.User.select():
+            user.tikets = 0
+            user.save()
+        for user in models.Ref.select():
+            user.msg_count = 0
+            user.save()
+        models.db.drop_tables([models.Tickets])
+        models.db.close()
+        bot.send_message(message.chat.id, "Все билеты были удалены")
+
+@bot.message_handler(commands=['stop'])
+def stop(message):
+    global stop
+    user_id = message.from_user.id
+    user_username = message.from_user.username
+    verification = config.is_admin(user_username, user_id)
+    if verification[0]:
+        stop = True
+        bot.send_message(message.chat.id, "Лотерея остановлена")
+
+@bot.message_handler(commands=['start_lottery'])
+def start(message):
+    global stop
+    user_id = message.from_user.id
+    user_username = message.from_user.username
+    verification = config.is_admin(user_username, user_id)
+    if verification[0]:
+        stop = False
+        bot.send_message(message.chat.id, "Лотерея запущена")
+
+@bot.message_handler(commands=['tickets'])
+def tickets(message):
+    user_id = message.from_user.id
+    user_username = message.from_user.username
+    verification = config.is_admin(user_username, user_id)
+    if verification[0] and len(message.text.split()) == 2:
+        params = list(message.text.split())
+        if len(params) == 2:
+            try:
+                models.db.connect()
+                user = models.User.get(models.User.user_id == params[1])
+                bot.send_message(message.chat.id, f"У пользователя {params[1]} {user.tikets} билетов")
+                models.db.close()
+            except:
+                bot.send_message(message.chat.id, "Пользователь не найден")
+        else:
+            bot.send_message(message.chat.id, "Неверный формат команды")
+    else:
+        models.db.connect()
+        user = models.User.get(models.User.user_id == user_id)
+        bot.send_message(message.chat.id, f"У Вас {user.tikets} билетов")
+        models.db.close()
+
+@bot.message_handler(commands=['delete'])
+def delete(message):
+    user_id = message.from_user.id
+    user_username = message.from_user.username
+    verification = config.is_admin(user_username, user_id)
+    if verification[0]:
+        params = list(message.text.split())
+        if len(params) == 2:
+                try:
+                    models.db.connect()
+                    user = models.User.get(models.User.nickname == params[1])
+                    bot.send_message(user.user_id, "Ваш аккаунт был удален")
+                    models.Tickets.delete().where(models.Tickets.user == user).execute()
+                    models.Ref.delete().where((models.Ref.invite_id == user) | (models.Ref.join_id == user)).execute()
+                    user.delete_instance()
+                    models.db.close()
+                    bot.send_message(message.chat.id, f"Пользователь {params[1]} был удален")
+                except:
+                    bot.send_message(message.chat.id, "Пользователь не найден")
+        else:
+            bot.send_message(message.chat.id, "Неверный формат команды")
 
 def is_user_subscribed(user_id) -> bool:
     channel_username = "@puton4ick"
@@ -114,49 +206,42 @@ def callback_inline(call: types.CallbackQuery):
 
 @bot.message_handler(func=lambda message: True)
 def count_messages(message):
+    global stop
     user_id = message.from_user.id
-    if len(message.text.split()) >= 3:
-        models.db.connect()
-        if models.User.select().where(models.User.user_id == user_id).exists():
-            user = models.User.get(models.User.user_id == user_id)
-            user.tikets += 1
-            user.save()
-            ticket = models.Tickets.create(user=user)
-            print(f"Ticket {ticket.id} created for user {user_id}")
-            if models.Ref.select().where(models.Ref.join_id == user_id).exists():
-                ref = models.Ref.get(models.Ref.join_id == user_id)
-                if ref.msg_count + 1 == 5:
-                    # Получаем пользователя, который присоединился по реферальной ссылке
-                    joined_user = models.User.get(models.User.user_id == user_id)
-                    # Увеличиваем количество его билетов на 5
-                    joined_user.tikets += 5
-                    joined_user.save()
+    if not stop:
+        if len(message.text.split()) >= 3:
+            models.db.connect()
+            if models.User.select().where(models.User.user_id == user_id).exists():
+                user = models.User.get(models.User.user_id == user_id)
+                user.tikets += 1
+                user.save()
+                ticket = models.Tickets.create(user=user)
+                print(f"Ticket {ticket.id} created for user {user_id}")
+                if models.Ref.select().where(models.Ref.join_id == user_id).exists():
+                    ref = models.Ref.get(models.Ref.join_id == user_id)
+                    if ref.msg_count + 1 == 5:
+                        joined_user = models.User.get(models.User.user_id == user_id)
+                        joined_user.tikets += 5
+                        joined_user.save()
 
-                    # Создаем 5 билетов для этого пользователя
-                    for _ in range(5):
-                        models.Tickets.create(user=joined_user)
+                        for _ in range(5):
+                            models.Tickets.create(user=joined_user)
 
-                    # Отправляем сообщение пользователю о получении билетов
-                    bot.send_message(joined_user.user_id, "Поздравляем! Вы получили 5 билетов!")
+                        bot.send_message(joined_user.user_id, "Поздравляем! Вы получили 5 билетов!")
 
-                    # Получаем пользователя, который поделился реферальной ссылкой
-                    invited_user = models.User.get(models.User.user_id == ref.invite_id)
-                    # Увеличиваем количество его билетов на 5
-                    invited_user.tikets += 5
-                    invited_user.save()
+                        invited_user = models.User.get(models.User.user_id == ref.invite_id)
+                        invited_user.tikets += 5
+                        invited_user.save()
 
-                    # Создаем 5 билетов для этого пользователя
-                    for _ in range(5):
-                        models.Tickets.create(user=invited_user)
+                        for _ in range(5):
+                            models.Tickets.create(user=invited_user)
 
-                    # Отправляем сообщение пользователю о получении билетов
-                    bot.send_message(invited_user.user_id, "Поздравляем! Вы получили 5 билетов!")
+                        bot.send_message(invited_user.user_id, "Поздравляем! Вы получили 5 билетов!")
 
-                    # Удаляем запись из таблицы Ref
-                    ref.delete_instance()
-                ref.msg_count += 1
-                ref.save()
-        models.db.close()
+                        ref.delete_instance()
+                    ref.msg_count += 1
+                    ref.save()
+            models.db.close()
 
 
 bot.polling(none_stop=True)
